typo
git-svn-id: svn://svn.dd-wrt.com/DD-WRT@29307 52c4871e-980c-0410-b1e0-e73912ce01f8

+++ src/router/httpd/visuals/wireless_ralink.c
@@ -396,7 +396,7 @@ void ej_active_wireless(webs_t wp, int argc, char_t ** argv)
 
 }
 
-extern float wifi_getrate(char *ifname);
+extern int wifi_getrate(char *ifname);
 
 #define KILO	1e3
 #define MEGA	1e6
